// src/app/modules/result/models/subject-template-master.ts
export class SubjectTemplateMaster {
    SUBJECT_TEMPLATE_ID: number;
    ORG_CODE: string;
    COURSE_OUTLINE_TEMPLATE_DESC: any;
    COMPULSORY_SUBJECT_ID_LIST: any;
    THIRD_SUBJECT_ID_LIST: any;
    FOURTH_SUBJECT_ID_LIST: any;
    RowStatus: number;
    User_Name: string;
    select: number;
}